IKASAN-2135 final changes for job lock management work.

+++ ikasaneip/spec/service/scheduled/src/main/java/org/ikasan/spec/scheduled/context/model/JobLockHolder.java
@@ -30,14 +30,14 @@ public interface JobLockHolder extends Serializable {
      *
      * @return
      */
-    long getLockCount();
+    int getLockCount();
 
     /**
      * Set the lock count
      *
      * @param lockCount
      */
-    void setLockCount(long lockCount);
+    void setLockCount(int lockCount);
 
     /**
      * Get the jobs associated with the lock
